Install all-the-icons in gitpod prebuild
Learned how to use emacs batch mode to run scripts noninteractively with
emacs. But to my surprise, successfully installing all-the-icons in
batch mode did not translate to gui emacs. I had to open a full emacs
and close it upon completion. This was the first version I wrote that
worked.